Rearrange querycoord trigger task code
